Changed Wrappers architecture
Moved a lot of methods from OrderWrapper and UnitWrapper into Order and Unit respectively, and turned them into public static functions with a Wrapper as an argument and referencing private/protected instance methods which contain the actual code for the function. This allows to greatly shrink the amount of code in the wrappers but still maintain decoupling, using the fact that private instance attributes and methods of an object can be accessed in a static function defined in that object's class.
Some stuff still had to be kept in UnitWrapper because of the need for polymorphism, as some methods implemented from IOrderable had still to be implemented there. As a matter of fact, there is no such thing as a "static interface methods".