Added a Build order prototype
This Build order takes a BuildingMarker, and if close enough to it, adds "HP" to it until it reaches its max HP or if the building unit runs out of "materials"
When the unit actually starts "building", a BuiltStructure is created with it associated marker wrapper set to the one the order was given.
If the order ends, the marker is destroyed if there's no actual BuiltStructure which has that marker, or if there is one but its HP is zero.

Now, how to deal with "Build at remote position" orders ? (i.e. the "real" build orders the player will issue)
Some testing and figuring out is still to be done.
Two possibilities for now :
- incorporating a simple MoveOrder and a simple BuildOrder into one big "MegaOrder", which will then "switch" from the first to the second, deciding on that using its own BehaviourTree just as a "normal" order would
- (prefered possibility) When the player issues a "build" order, two orders should be created in actuality : a move one (requires waypoints) and a simple build order (requires build marker)