Attempt at another representation for order "plans" for units. (see UnitWrapper)
There are "active" orders, and "passive" orders.
Two LinkedLists represent them.
Active ones correspond to the "main" orders, like movement, building...
Only one active order can run at a time.
Passive ones correspond to "secondary" orders mostly acting as "rules" and implementing behaviour specific to those rules, in a way. (Which makes their purpose different from just "order parameter tags" (OrderParams))
One (and only for now) example in this commit : EngageAtPositionsOrder. It tells a unit to watch / engage targets that are only in the specfied "circles" (FireAtPositionMarker). It's a effectively part of its Rules Of Engagement (ROE).
And as many passive orders can run at a time.

More testing and figuring out still must be done to see how to find a balance between "customization" / possibilities for order behaviour and complexity, which must not become too important and useless.
For now, the main idea is to implement "basic" functions in UnitWrapper which may be used in OrderWrappers to specify behaviour and reactions to other orders - both active and passive ! - typically at the end or at the the start of an order. Also ExecutionModes in OrderParams can help to establish most common ways of interaction. Until refactoring is needed and I have an idea whether it was a good or bad idea, both in terms of possible gameplay and programming complexity.
Currently, it is possible for an order to insert another one right before it, and also, when finished to start the next active order when it's execution mode includes "Chain".
Other execution modes are in preparation : NotifyAtEnd, AskConfirmationBeforeStart...